refactor(runbotics-orchestrator-ui): :art: Fix typos in naming in credentials folder and getDownloadCloudFileAction func
